+++ R/across.R
@@ -2,8 +2,9 @@
 #'
 #' `across()` makes it easy to apply the same transformation to multiple
 #' columns, allowing you to use [select()] semantics inside in [summarise()] and
-#' [mutate()]. `across()` replaces the family of "scoped variants" like
+#' [mutate()]. `across()` supersedes the family of "scoped variants" like
 #' `summarise_at()`, `summarise_if()`, and `summarise_all()`.
+#' See `vignette("colwise")` for more details.
 #'
 #' @param cols <[`tidy-select`][dplyr_tidy_select]> Columns to transform.
 #'   Because `across()` is used within functions like `summarise()` and