+++ R/colwise-arrange.R
@@ -1,5 +1,11 @@
 #' Arrange rows by a selection of variables
 #'
+#' @description
+#' \Sexpr[results=rd, stage=render]{lifecycle::badge("retired")}
+#'
+#' Scoped verbs (`_if`, `_at`, `_all`) have been superseded by the use of
+#' [across()] in an existing verb. See `vignette("colwise")` for details.
+#'
 #' These [scoped] variants of [arrange()] sort a data frame by a
 #' selection of variables. Like [arrange()], you can modify the
 #' variables before ordering with the `.funs` argument.
@@ -15,14 +21,13 @@
 #' @export
 #' @examples
 #' df <- as_tibble(mtcars)
-#' df
 #' arrange_all(df)
+#' # ->
+#' arrange(df, across())
 #'
-#' # You can supply a function that will be applied before taking the
-#' # ordering of the variables. The variables of the sorted tibble
-#' # keep their original values.
 #' arrange_all(df, desc)
-#' arrange_all(df, list(~desc(.)))
+#' # ->
+#' arrange(df, across(everything(), desc))
 arrange_all <- function(.tbl, .funs = list(), ..., .by_group = FALSE) {
   funs <- manip_all(.tbl, .funs, enquo(.funs), caller_env(), .include_group_vars = TRUE, ...)
   if (!length(funs)) {