+++ R/filter.R
@@ -1,10 +1,10 @@
-#' Return rows with matching conditions
+#' Subset rows using column values
 #'
-#' Use `filter()` to choose rows/cases where conditions are true. Unlike
-#' base subsetting with `[`, rows where the condition evaluates to `NA` are
-#' dropped.
+#' `filter()` retains the rows where the conditions you provide a `TRUE`. Note
+#' that, unlike base subsetting with `[`, rows where the condition evaluates
+#' to `NA` are dropped.
 #'
-#' Note that dplyr is not yet smart enough to optimise filtering optimisation
+#' dplyr is not yet smart enough to optimise filtering optimisation
 #' on grouped datasets that don't need grouped calculations. For this reason,
 #' filtering is often considerably faster on [ungroup()]ed data.
 #'
@@ -34,32 +34,30 @@
 #'
 #' The former keeps rows with `mass` greater than the global average
 #' whereas the latter keeps rows with `mass` greater than the gender
+#'
 #' average.
-#'
-#' It is valid to use grouping variables in filter expressions.
-#'
-#' When applied on a grouped tibble, `filter()` automatically [rearranges][arrange]
-#' the tibble by groups for performance reasons.
-#'
-#' @section Tidy data:
-#' When applied to a data frame, row names are silently dropped. To preserve,
-#' convert to an explicit variable with [tibble::rownames_to_column()].
-#'
-#' @section Scoped filtering:
-#' The three [scoped] variants ([filter_all()], [filter_if()] and
-#' [filter_at()]) make it easy to apply a filtering condition to a
-#' selection of variables.
-#'
 #' @family single table verbs
-#' @param .data A tbl. All main verbs are S3 generics and provide methods
-#'   for [tbl_df()], [dtplyr::tbl_dt()] and [dbplyr::tbl_dbi()].
+#' @inheritParams arrange
 #' @param ... <[`tidy-eval`][dplyr_tidy_eval]> Logical predicates defined in
 #'   terms of the variables in `.data`.
 #'   Multiple conditions are combined with `&`. Only rows where the
 #'   condition evaluates to `TRUE` are kept.
 #' @param .preserve when `FALSE` (the default), the grouping structure
 #'   is recalculated based on the resulting data, otherwise it is kept as is.
-#' @inherit arrange return
+#' @return
+#' An object of the same type as `.data`.
+#'
+#' * Rows are a subset of the input, but appear in the same order.
+#' * Columns are not modified.
+#' * The number of groups may be reduced (if `.preserve` is not `TRUE`).
+#' * Data frame attributes are preserved.
+#' @section Methods:
+#' This function is a **generic**, which means that packages can provide
+#' implementations (methods) for other classes. See the documentation of
+#' individual methods for extra arguments and differences in behaviour.
+#'
+#' The following methods are currently available in loaded packages:
+#' \Sexpr[stage=render,results=rd]{dplyr:::methods_rd("filter")}.
 #' @seealso [filter_all()], [filter_if()] and [filter_at()].
 #' @export
 #' @examples
@@ -105,24 +103,8 @@ filter.data.frame <- function(.data, ..., .preserve = FALSE) {
     return(.data)
   }
 
-  idx <- filter_rows(.data, ...)
-  .data[idx[[1]], , drop = FALSE]
-}
-
-#' @export
-filter.grouped_df <- function(.data, ..., .preserve = !group_by_drop_default(.data)) {
-  if (missing(...)) {
-    return(.data)
-  }
-
-  idx <- filter_rows(.data, ...)
-  data <- as.data.frame(.data)[idx[[1]], , drop = FALSE]
-
-  groups <- group_data(.data)
-  groups$.rows <- filter_update_rows(nrow(.data), idx[[3]], idx[[1]], idx[[2]])
-  groups <- group_data_trim(groups, .preserve)
-
-  new_grouped_df(data, groups)
+  loc <- filter_rows(.data, ...)[[1]]
+  dplyr_row_slice(.data, loc, preserve = .preserve)
 }
 
 filter_rows <- function(.data, ...) {
@@ -159,9 +141,4 @@ check_filter <- function(dots) {
     }
 
   }
-}
-
-
-filter_update_rows <- function(n_rows, group_indices, keep, new_rows_sizes) {
-  .Call(`dplyr_filter_update_rows`, n_rows, group_indices, keep, new_rows_sizes)
 }