+++ R/join-rows.R
@@ -2,9 +2,9 @@ join_rows <- function(x_key, y_key, type = c("inner", "left", "right", "full")) 
   type <- arg_match(type)
 
   # Find matching rows in y for each row in x
-  y_split <- vec_group_pos(y_key)
+  y_split <- vec_group_loc(y_key)
   matches <- vec_match(x_key, y_split$key)
-  y_loc <- y_split$pos[matches]
+  y_loc <- y_split$loc[matches]
 
   if (type == "left" || type == "full") {
     y_loc <- map(y_loc, function(x) if (is.null(x)) NA_integer_ else x)