+++ R/mutate.R
@@ -1,10 +1,11 @@
-#' Create or transform variables
+#' Create, modify, and delete columns
 #'
 #' `mutate()` adds new variables and preserves existing ones;
 #' `transmute()` adds new variables and drops existing ones.
 #' New variables overwrite existing variables of the same name.
+#' Variables can be removed by setting their value to `NULL`.
 #'
-#' @section Useful functions available in calculations of variables:
+#' @section Useful mutate functions:
 #'
 #' * [`+`], [`-`], [log()], etc., for their usual mathematical meanings
 #'
@@ -44,31 +45,49 @@
 #' The former normalises `mass` by the global average whereas the
 #' latter normalises by the averages within gender levels.
 #'
-#' `mutate()` does not evaluate the expressions when the group is empty.
-#'
-#' @section Scoped mutation and transmutation:
-#'
-#' The three [scoped] variants of `mutate()` ([mutate_all()],
-#' [mutate_if()] and [mutate_at()]) and the three variants of
-#' `transmute()` ([transmute_all()], [transmute_if()],
-#' [transmute_at()]) make it easy to apply a transformation to a
-#' selection of variables.
-#'
 #' @export
-#' @inheritParams filter
-#' @inheritSection filter Tidy data
-#' @param ... <[`tidy-eval`][dplyr_tidy_eval]> Name-value pairs of expressions,
-#'   each with length 1 or the same length as the number of rows in the group
-#'   (if using [group_by()]) or in the entire input (if not using groups).
-#'   The name of each argument will be the name of a new variable, and the
-#'   value will be its corresponding value. Use a `NULL` value in `mutate`
-#'   to drop a variable.  New variables overwrite existing variables
-#'   of the same name.
+#' @inheritParams arrange
+#' @param ... <[`tidy-eval`][dplyr_tidy_eval]> Name-value pairs.
+#'   The name gives the name of the column in the output.
+#'
+#'   The value can be:
+#'
+#'   * A vector of length 1, which will be recycled to the correct length.
+#'   * A vector the same length as the current group (or the whole data frame
+#'     if ungrouped).
+#'   * `NULL`, to remove the column.
+#'   * A data frame or tibble, to create multiple columns in the output.
+#'
 #' @family single table verbs
 #' @return
-#' An object of the same type as `.data`. The rows will be left as; only
-#' the columns will be changed. `mutate()` preserves order of existing columns,
-#' adding new columns to the right. `transmute()` drops existing columns.
+#' An object of the same type as `.data`.
+#'
+#' For `mutate()`:
+#'
+#' * Rows are not affected.
+#' * Existing columns will be preserved unless explicitly modified.
+#' * New columns will be added to the right of existing columns.
+#' * Columns given value `NULL` will be removed
+#' * Groups will be recomputed if a grouping variable is mutated.
+#' * Data frame attributes are preserved.
+#'
+#' For `transmute()`:
+#'
+#' * Rows are not affected.
+#' * Apart from grouping variables, existing columns will be remove unless
+#'   explicitly kept.
+#' * Column order matches order of expressions.
+#' * Groups will be recomputed if a grouping variable is mutated.
+#' * Data frame attributes are preserved.
+#' @section Methods:
+#' These function are **generic**s, which means that packages can provide
+#' implementations (methods) for other classes. See the documentation of
+#' individual methods for extra arguments and differences in behaviour.
+#'
+#' Methods available in currently loaded packages:
+#'
+#' * `mutate()`: \Sexpr[stage=render,results=rd]{dplyr:::methods_rd("mutate")}.
+#' * `transmute()`: \Sexpr[stage=render,results=rd]{dplyr:::methods_rd("transmute")}.
 #' @examples
 #' # Newly created variables are available immediately
 #' mtcars %>% as_tibble() %>% mutate(
@@ -76,22 +95,18 @@
 #'   cyl4 = cyl2 * 2
 #' )
 #'
-#' # You can also use mutate() to remove variables and
-#' # modify existing variables
+#' # As well as adding new variables, you can use mutate() to
+#' # remove variables and modify existing variables.
 #' mtcars %>% as_tibble() %>% mutate(
 #'   mpg = NULL,
 #'   disp = disp * 0.0163871 # convert to litres
 #' )
-#'
 #'
 #' # window functions are useful for grouped mutates
 #' mtcars %>%
 #'  group_by(cyl) %>%
 #'  mutate(rank = min_rank(desc(mpg)))
 #' # see `vignette("window-functions")` for more details
-#'
-#' # You can drop variables by setting them to NULL
-#' mtcars %>% mutate(cyl = NULL)
 #'
 #' # mutate() vs transmute --------------------------
 #' # mutate() keeps all existing variables
@@ -102,7 +117,7 @@
 #' mtcars %>%
 #'   transmute(displ_l = disp / 61.0237)
 #'
-#'
+#' # Grouping ----------------------------------------
 #' # The mutate operation may yield different results on grouped
 #' # tibbles because the expressions are computed within groups.
 #' # The following normalises `mass` by the global average:
@@ -117,11 +132,7 @@
 #'   mutate(mass / mean(mass, na.rm = TRUE)) %>%
 #'   pull()
 #'
-#' # Note that you can't overwrite grouping variables:
-#' gdf <- mtcars %>% group_by(cyl)
-#' try(mutate(gdf, cyl = cyl * 100))
-#'
-#'
+#' # Indirection ----------------------------------------
 #' # Refer to column names stored as strings with the `.data` pronoun:
 #' vars <- c("mass", "height")
 #' mutate(starwars, prod = .data[[vars[[1]]]] * .data[[vars[[2]]]])
@@ -133,33 +144,7 @@ mutate <- function(.data, ...) {
 #' @export
 mutate.data.frame <- function(.data, ...) {
   cols <- mutate_cols(.data, ...)
-  if (is.null(cols)) {
-    return(.data)
-  }
-
-  .data <- .data[setdiff(names(.data), cols$delete)]
-  .data[names(cols$add)] <- cols$add
-
-  .data
-}
-
-#' @export
-mutate.grouped_df <- function(.data, ...) {
-  cols <- mutate_cols(.data, ...)
-  if (is.null(cols)) {
-    return(.data)
-  }
-
-  data <- as_tibble(.data)
-  data <- data[setdiff(names(.data), cols$delete)]
-  data[names(cols$add)] <- cols$add
-
-  groups <- group_data(.data)
-  if (any(cols$change %in% names(groups))) {
-    grouped_df(data, group_vars(.data), group_by_drop_default(.data))
-  } else {
-    new_grouped_df(data, groups)
-  }
+  dplyr_col_modify(.data, cols)
 }
 
 #' @rdname mutate
@@ -171,33 +156,16 @@ transmute <- function(.data, ...) {
 #' @export
 transmute.data.frame <- function(.data, ...) {
   cols <- mutate_cols(.data, ...)
-  if (is.null(cols)) {
-    return(.data[integer()])
-  }
+  .data <- dplyr_col_modify(.data, cols)
 
-  .data <- .data[integer()]
-  .data[names(cols$add)] <- cols$add
+  out_cols <- c(
+    # ensure group vars present
+    setdiff(group_vars(.data), names(cols)),
+    # cols might contain NULLs
+    intersect(names(cols), names(.data))
+  )
 
-  .data
-}
-
-#' @export
-transmute.grouped_df <- function(.data, ...) {
-  cols <- mutate_cols(.data, ...)
-  if (is.null(cols)) {
-    return(.data[group_vars(.data)])
-  }
-
-  data <- as_tibble(.data)
-  data <- data[setdiff(group_vars(.data), names(cols$add))]
-  data[names(cols$add)] <- cols$add
-
-  groups <- group_data(.data)
-  if (any(cols$change %in% names(groups))) {
-    grouped_df(data, group_vars(.data), group_by_drop_default(.data))
-  } else {
-    new_grouped_df(data, groups)
-  }
+  .data[out_cols]
 }
 
 # Helpers -----------------------------------------------------------------
@@ -222,56 +190,74 @@ mutate_cols <- function(.data, ...) {
 
   new_columns <- list()
 
-  for (i in seq_along(dots)) {
-    # a list in which each element is the result of
-    # evaluating the quosure in the "sliced data mask"
-    # recycling it appropriately to match the group size
-    #
-    # TODO: reinject hybrid evaluation at the R level
-    c(chunks, needs_recycle) %<-% mask$eval_all_mutate(dots[[i]], dots_names, i)
+  tryCatch({
+    for (i in seq_along(dots)) {
+      # a list in which each element is the result of
+      # evaluating the quosure in the "sliced data mask"
+      # recycling it appropriately to match the group size
+      #
+      # TODO: reinject hybrid evaluation at the R level
+      c(chunks, needs_recycle) %<-% mask$eval_all_mutate(dots[[i]])
 
-    if (is.null(chunks)) {
-      if (!is.null(dots_names) && dots_names[i] != "") {
-        new_columns[[dots_names[i]]] <- zap()
+      if (is.null(chunks)) {
+        if (!is.null(dots_names) && dots_names[i] != "") {
+          new_columns[[dots_names[i]]] <- zap()
 
-        # we might get a warning if dots_names[i] does not exist
-        suppressWarnings(mask$remove(dots_names[i]))
+          # we might get a warning if dots_names[i] does not exist
+          suppressWarnings(mask$remove(dots_names[i]))
+        }
+        next
       }
-      next
+
+      if (needs_recycle) {
+        chunks <- map2(chunks, rows_lengths, function(chunk, n) {
+          vec_recycle(chunk, n)
+        })
+      }
+      result <- vec_slice(vec_c(!!!chunks), o_rows)
+
+      not_named <- (is.null(dots_names) || dots_names[i] == "")
+      if (not_named && is.data.frame(result)) {
+        new_columns[names(result)] <- result
+
+        # remember each result separately
+        map2(seq_along(result), names(result), function(i, nm) {
+          mask$add(nm, pluck(chunks, i))
+        })
+      } else {
+        name <- if (not_named) auto_named_dots[i] else dots_names[i]
+
+        # treat as a single output otherwise
+        new_columns[[name]] <- result
+
+        # remember
+        mask$add(name, chunks)
+      }
+
     }
 
-    if (needs_recycle) {
-      chunks <- map2(chunks, rows_lengths, function(chunk, n) {
-        vec_recycle(chunk, n)
-      })
+  },
+    rlang_error_data_pronoun_not_found = function(e) {
+      stop_error_data_pronoun_not_found(conditionMessage(e), index = i, dots = dots, fn = "mutate")
+    },
+    vctrs_error_recycle_incompatible_size = function(e) {
+      stop_mutate_recycle_incompatible_size(e, index = i, dots = dots)
+    },
+    dplyr_mutate_mixed_NULL = function(e) {
+      stop_mutate_mixed_NULL(index = i, dots = dots)
+    },
+    dplyr_mutate_not_vector = function(e) {
+      stop_mutate_not_vector(index = i, dots = dots, result = e$result)
+    },
+    vctrs_error_incompatible_type = function(e) {
+      stop_combine(conditionMessage(e), index = i, dots = dots, fn = "mutate")
+    },
+    simpleError = function(e) {
+      stop_eval_tidy(e, index = i, dots = dots, fn = "mutate")
     }
-    result <- vec_slice(vec_c(!!!chunks), o_rows)
-
-    not_named <- (is.null(dots_names) || dots_names[i] == "")
-    if (not_named && is.data.frame(result)) {
-      new_columns[names(result)] <- result
-
-      # remember each result separately
-      map2(seq_along(result), names(result), function(i, nm) {
-        mask$add(nm, pluck(chunks, i))
-      })
-    } else {
-      name <- if (not_named) auto_named_dots[i] else dots_names[i]
-
-      # treat as a single output otherwise
-      new_columns[[name]] <- result
-
-      # remember
-      mask$add(name, chunks)
-    }
-
-  }
+  )
 
   is_zap <- map_lgl(new_columns, inherits, "rlang_zap")
-
-  list(
-    add = new_columns[!is_zap],
-    delete = names(new_columns)[is_zap],
-    change = names(new_columns)
-  )
+  new_columns[is_zap] <- rep(list(NULL), sum(is_zap))
+  new_columns
 }