+++ tests/testthat/test-transmute.R
@@ -14,6 +14,19 @@ test_that("transmute preserves variable of output", {
   expect_named(gf %>% transmute(y = 1, x = 2), c("y", "x"))
 })
 
+test_that("transmute preserves grouping", {
+  gf <- group_by(tibble(x = 1:2, y = 2), x)
+
+  i <- count_regroups(out <- transmute(gf, x = 1))
+  expect_equal(i, 1L)
+  expect_equal(group_vars(out), "x")
+  expect_equal(nrow(group_data(out)), 1)
+
+  i <- count_regroups(out <- transmute(gf, z = 1))
+  expect_equal(i, 0)
+  expect_equal(group_data(out), group_data(gf))
+})
+
 # Empty transmutes -------------------------------------------------
 
 test_that("transmute with no args returns grouping vars", {